requirements.txt
Flask==1.0.2
google-cloud-datastore==1.7.3
numpy==1.10.1
